+++ extensions/upstash/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Upstash Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version - Redis + Vector] - {PR_MERGE_DATE}
+
+- List Redis Databases
+    - View Details & Usage
+    - Create Database
+- List Vector Indices
+    - Create Index
+    - Delete Index